WIP: replace ws with socket.io
