+++ packages/frontend/src/contexts/SocketContext.tsx
@@ -1,6 +1,7 @@
 import { createContext } from "react";
 
 const SocketContext = createContext({
+  //examples only
   queueLength: 0,
   positionInLine: 0
 });
