+++ packages/frontend/src/contexts/sockets/emit.tsx
@@ -1,10 +1,5 @@
 import { socket } from "./index";
-export const addClientToQueue = () => {
-  socket.emit("addClientIdToQueue");
-};
-export const getQueueLength = () => {
-  socket.emit("queueLengthToSocket");
-};
-export const removeUserFromQueue = () => {
-  socket.emit("removeUserFromQueue");
+
+export const henloServer = (message: string) => {
+  socket.emit("henloServer", message);
 };
